flex:fn long double to double

+++ src/flex/fn/fn.c
@@ -27,11 +27,11 @@ getsym (char const *name)
   return NULL;
 }
 
-long double
-fact (long double d)
+double
+fact (double d)
 {
   long n = (long)d;
-  long double acc = 1;
+  double acc = 1;
   while (n > 1)
     {
       acc *= n--;
